+++ llm_infer/serving/dispatch/handlers/__init__.py
@@ -2,10 +2,12 @@
 
 from .batching import ContinuousBatchingHandler
 from .bounded import BoundedQueueHandler
+from .concurrent_http import ConcurrentHttpHandler
 from .sequential import SequentialHandler
 
 __all__ = [
     "SequentialHandler",
     "BoundedQueueHandler",
     "ContinuousBatchingHandler",
+    "ConcurrentHttpHandler",
 ]